se agregó la función formatearHora que trae la hora del pedido en el administradorPedidos. y se agregó un Item de la Hora en el Acordeón del pedido.
